requirements.txt
accelerate==0.18.0
stability-sdk==0.7.0
streamlit==1.22.0
streamlit-drawable-canvas==0.9.2
diffusers==0.16.1
huggingface-hub==0.14.1
imageio==2.27.0
importlib-metadata==6.6.0
numpy
Pillow==9.5.0
scipy==1.10.1
sympy==1.11.1
tenacity==8.2.2
torch==2.0.0
transformers==4.28.1
safetensors==0.3.1
